Documentação das views e correção de bugs

+++ album/views.py
@@ -11,12 +11,14 @@ from . import models
 
 
 class Album(ListView):
+    """ Page with all published albuns with 3 images or more """
     template_name = 'album/album.html'
     model = models.Album
     paginate_by = 6
     context_object_name = 'albuns'
 
     def get_queryset(self, *args, **kwargs):
+        """ Select the published albuns and check if the album has 3 images or more"""
         qs = super().get_queryset(*args, **kwargs)
 
         qs = qs.filter(published_album=True)
@@ -27,6 +29,7 @@ class Album(ListView):
         return qs
 
     def get_context_data(self, *args, **kwargs):
+        """ Add the images for each album """
         context = super().get_context_data(*args, **kwargs)
 
         albuns_images = list()
@@ -39,16 +42,19 @@ class Album(ListView):
 
 
 class AlbumImages(DetailView):
+    """ Shows all images of the selected album """
     template_name = 'album/images.html'
     model = models.Album
     context_object_name = 'album'
 
     def get_context_data(self, *args, **kwargs):
+        """ Select the images of the album """
         context = super().get_context_data(*args, **kwargs)
         context['images'] = models.Image.objects.filter(album_image=context.get('album').pk)
         return context
 
     def get(self, request, *args, **kwargs):
+        """ Make sure the album has been published and the number of images is 3 or more """
         self.object = self.get_object()
         context = self.get_context_data(object=self.object)
 
@@ -62,6 +68,7 @@ class AlbumImages(DetailView):
 
 
 class AlbumUser(LoginRequiredMixin, PermissionRequiredMixin, ListView):
+    """ Shows all the user albuns """
     template_name = 'album/album_user.html'
     model = models.Album
     paginate_by = 10
@@ -71,17 +78,20 @@ class AlbumUser(LoginRequiredMixin, PermissionRequiredMixin, ListView):
     permission_denied_message = 'Necessário usuário autorizado'
 
     def get_queryset(self, *args, **kwargs):
+        """ Select only the user albuns """
         qs = super().get_queryset(*args, **kwargs)
         qs = qs.filter(user_album=self.request.user.pk)
         return qs
 
     def get_context_data(self, *args, **kwargs):
+        """ Add the album form to the context """
         context = super().get_context_data(*args, **kwargs)
         context['album_form'] = forms.AlbumForm(self.request.POST or None)
         return context
 
 
 class AlbumCreate(LoginRequiredMixin, PermissionRequiredMixin, CreateView):
+    """ Page to create a new album """
     template_name = 'album/album_create.html'
     model = models.Album
     form_class = forms.AlbumForm
@@ -90,6 +100,7 @@ class AlbumCreate(LoginRequiredMixin, PermissionRequiredMixin, CreateView):
     permission_denied_message = 'Necessário usuário autorizado'
 
     def post(self, request, *args, **kwargs):
+        """ Make sure the form is valid e add the required fields """
         form = self.get_form()
         if form.is_valid():
             form = form.save(commit=False)
@@ -99,6 +110,7 @@ class AlbumCreate(LoginRequiredMixin, PermissionRequiredMixin, CreateView):
             return self.form_invalid(form)
 
     def get_success_url(self):
+        """ Success message """
         messages.success(self.request, 'Álbum cadastrado')
         return reverse('album:user_album')
 
@@ -106,40 +118,49 @@ class AlbumCreate(LoginRequiredMixin, PermissionRequiredMixin, CreateView):
 @login_required(login_url=settings.LOGIN_URL)
 @permission_required(perm='album.change_album')
 def album_update(request):
-    pk = request.POST.get('primary-key')
-    form = forms.AlbumForm(request.POST)
-    if pk is not None:
-        album = get_object_or_404(models.Album, pk=pk, user_album=request.user)
-        if form.is_valid():
-            album_form = form.save(commit=False)
-            album_form.pk = album.pk
-            album_form.user_album = album.user_album
-            album_form.save()
-            messages.success(request, 'Álbum editado')
-            return redirect(reverse('album:user_album'))
+    """ Method to update the selected album """
+    if request.POST:
+        pk = request.POST.get('primary-key')
+        form = forms.AlbumForm(request.POST)
+        if pk is not None:
+            album = get_object_or_404(models.Album, pk=pk, user_album=request.user)
+            if form.is_valid():
+                album_form = form.save(commit=False)
+                album_form.pk = album.pk
+                album_form.user_album = album.user_album
+                album_form.save()
+                messages.success(request, 'Álbum editado')
+                return redirect(reverse('album:user_album'))
+            else:
+                messages.error(request, 'Dados incorretos')
+                return redirect(reverse('album:user_album'))
         else:
-            messages.error(request, 'Dados incorretos')
+            messages.error(request, 'Álbum não encontrado')
             return redirect(reverse('album:user_album'))
     else:
-        messages.error(request, 'Álbum não encontrado')
         return redirect(reverse('album:user_album'))
 
 
 @login_required(login_url=settings.LOGIN_URL)
 @permission_required(perm='album.delete_album')
 def album_delete(request):
-    pk = request.POST.get('primary-key')
-    if pk is not None:
-        album = get_object_or_404(models.Album, pk=pk, user_album=request.user)
-        album.delete()
-        messages.error(request, 'Álbum deletado')
-        return redirect(reverse('album:user_album'))
+    """ Method to delete the selected album """
+    if request.POST:
+        pk = request.POST.get('primary-key')
+        if pk is not None:
+            album = get_object_or_404(models.Album, pk=pk, user_album=request.user)
+            album.delete()
+            messages.error(request, 'Álbum deletado')
+            return redirect(reverse('album:user_album'))
+        else:
+            messages.error(request, 'Álbum não encontrado')
+            return redirect(reverse('album:user_album'))
     else:
-        messages.error(request, 'Álbum não encontrado')
         return redirect(reverse('album:user_album'))
 
 
 class AlbumImagesUser(LoginRequiredMixin, PermissionRequiredMixin, DetailView):
+    """ Page with all the images of user album """
     template_name = 'album/album_images.html'
     model = models.Album
     context_object_name = 'album'
@@ -148,6 +169,7 @@ class AlbumImagesUser(LoginRequiredMixin, PermissionRequiredMixin, DetailView):
     permission_denied_message = 'Necessário usuário autorizado'
 
     def get_context_data(self, *args, **kwargs):
+        """ Add all the images to the context and the form to add new images """
         context = super().get_context_data(*args, **kwargs)
 
         context['images'] = models.Image.objects.filter(album_image=context.get('album').pk)
@@ -157,12 +179,14 @@ class AlbumImagesUser(LoginRequiredMixin, PermissionRequiredMixin, DetailView):
 
 
 class UploadImages(LoginRequiredMixin, PermissionRequiredMixin, View, FormMixin):
+    """ Page to view the images of the user album and upload new images """
     form_class = forms.ImageForm
     login_url = settings.LOGIN_URL
     permission_required = 'album.add_image'
     permission_denied_message = 'Necessário usuário autorizado'
 
     def post(self, request, *args, **kwargs):
+        """ Upload the selected images """
         form = self.get_form()
         album = get_object_or_404(models.Album, pk=kwargs.get('pk'), user_album=request.user.pk)
         if form.is_valid():
@@ -180,38 +204,50 @@ class UploadImages(LoginRequiredMixin, PermissionRequiredMixin, View, FormMixin)
 @login_required(login_url=settings.LOGIN_URL)
 @permission_required(perm='album.change_image')
 def image_update_title(request, pk):
+    """ Update the image title """
     album = get_object_or_404(models.Album, pk=pk, user_album=request.user.pk)
-    image_pk = request.POST.get('primary-key')
-    title = request.POST.get('title_image')
-    if len(title) < 5:
-        messages.error(request, 'Título deve possuir ao menos 5 caracteres')
-        return redirect(reverse('album:user_images', args=[album.pk, ]))
+    if request.POST:
+        image_pk = request.POST.get('primary-key')
+        title = request.POST.get('title_image')
+        if len(title) < 5:
+            messages.error(request, 'Título deve possuir ao menos 5 caracteres')
+            return redirect(reverse('album:user_images', args=[album.pk, ]))
 
-    image = get_object_or_404(models.Image, pk=image_pk, album_image=album.pk)
-    image.title_image = title
-    image.save()
-    messages.success(request, 'Título editado')
-    return redirect(reverse('album:user_images', args=[album.pk, ]))
+        image = get_object_or_404(models.Image, pk=image_pk, album_image=album.pk)
+        image.title_image = title
+        image.save()
+        messages.success(request, 'Título editado')
+        return redirect(reverse('album:user_images', args=[album.pk, ]))
+    else:
+        return redirect(reverse('album:user_images', args=[album.pk, ]))
 
 
 @login_required(login_url=settings.LOGIN_URL)
 @permission_required(perm='album.delete_image')
 def image_delete(request, pk):
+    """ Delete the selected image """
     album = get_object_or_404(models.Album, pk=pk, user_album=request.user)
-    image_pk = request.POST.get('primary-key')
-    image = get_object_or_404(models.Image, pk=image_pk, album_image=album.pk)
-    image.delete()
-    messages.success(request, 'Imagem deletada')
-    return redirect(reverse('album:user_images', args=[album.pk, ]))
+    if request.POST:
+        image_pk = request.POST.get('primary-key')
+        image = get_object_or_404(models.Image, pk=image_pk, album_image=album.pk)
+        image.delete()
+        messages.success(request, 'Imagem deletada')
+        return redirect(reverse('album:user_images', args=[album.pk, ]))
+    else:
+        return redirect(reverse('album:user_images', args=[album.pk, ]))
 
 
 @login_required(login_url=settings.LOGIN_URL)
 @permission_required(perm='album.delete_image')
 def multiple_image_delete(request, pk):
-    images_pk = request.POST.getlist('delete-items') or []
-    if len(images_pk):
-        album = get_object_or_404(models.Album, pk=pk, user_album=request.user.pk)
-        images = models.Image.objects.filter(pk__in=images_pk, album_image=album.pk)
-        images.delete()
-        messages.success(request, "Imagens excluidas")
-    return redirect(reverse('album:user_images', args=[pk, ]))
+    """ Method to delete multiple images at the same time """
+    if request.POST:
+        images_pk = request.POST.getlist('delete-items') or []
+        if len(images_pk):
+            album = get_object_or_404(models.Album, pk=pk, user_album=request.user.pk)
+            images = models.Image.objects.filter(pk__in=images_pk, album_image=album.pk)
+            images.delete()
+            messages.success(request, "Imagens excluidas")
+        return redirect(reverse('album:user_images', args=[pk, ]))
+    else:
+        return redirect(reverse('album:user_images', args=[pk, ]))
